refactor(tests): reorganize test structure to mirror source project refs #480
Reorganize test project to match source project folder structure:
- Move endpoint tests to Features/{Feature}/Endpoints/
- Move entity model tests to Features/Entities/Models/
- Move ImageConfigurationTests to Infrastructure/Storage/
- Update all namespaces to reflect new structure
- Remove empty top-level test directories
